requirements.txt
streamlit==1.52.2
requests==2.32.5
Pillow==12.1.0
numpy==2.4.0
pandas==2.3.3
urllib3==2.6.2
google-generativeai==0.7.2
